# api/openapi.yaml
openapi: 3.1.0
info:
  title: Clustr API
  version: 0.1.0
  description: Production messaging and shared-group API. Message and media content is client-encrypted.
servers:
  - url: https://api.clustr.example/v1
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      security: []
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
      responses:
        "200":
          description: Rotated token pair
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TokenPair"}
...
  /me:
    get:
      operationId: me
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: {$ref: "#/components/schemas/User"}
    patch:
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: true}
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: {$ref: "#/components/schemas/User"}
    delete:
      operationId: deleteAccount
      description: Irreversibly deletes authentication and discovery data, removes group memberships, and preserves shared financial history under a Deleted user tombstone.
      responses:
        "204": {description: Account deleted and all sessions revoked}
        "401": {description: Access token is missing, invalid, or already revoked}
...
  /devices/{device_id}:
    put:
      operationId: upsertDevice
      parameters:
        - {$ref: "#/components/parameters/DeviceID"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, platform]
              properties:
                name: {type: string}
                platform: {type: string}
                push_token: {type: string}
                identity_key: {type: string, description: Base64 public identity key}
                signed_prekey: {$ref: "#/components/schemas/SignedPreKey"}
      responses:
        "200":
          description: Device registered
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Device"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DeviceID: {name: device_id, in: path, required: true, schema: {type: string, format: uuid}}
...
  schemas:
...
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_at]
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        token_type: {type: string, const: Bearer}
        expires_at: {type: string, format: date-time}
...
    User:
      type: object
      required: [id, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        email: {type: string}
        phone: {type: string}
        display_name: {type: string}
        avatar_url: {type: string, format: uri}
        profile: {type: object, additionalProperties: true}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Device:
      type: object
      properties:
        id: {type: string, format: uuid}
        user_id: {type: string, format: uuid}
        name: {type: string}
        platform: {type: string}
        identity_key: {type: string}
        signed_prekey: {$ref: "#/components/schemas/SignedPreKey"}
    SignedPreKey:
      type: object
      required: [key_id, public_key, signature]
      additionalProperties: false
      properties:
        key_id: {type: integer, minimum: 0}
        public_key: {type: string, contentEncoding: base64}
        signature: {type: string, contentEncoding: base64}
...